+++ src/components/Alert/CHANGELOG.md
@@ -1,3 +1,4 @@
 - 2025-11-30: Initial version.
 - 2025-12-12: Update the styles - padding and gap - to make it look better.
 - 2025-12-13: Update to make it look good with new colors.
+- 2026-02-12: Update the styles to make it looks better.